requirements.txt
numpy==2.0.2
opencv-python==4.11.0.86
pillow==10.2.0
matplotlib==3.10.0
tensorflow==2.18.0